ヘッダに分割
元々のrotation.*をリネーム(それによりSample内も書き換え)
