+++ jissennohito-server/rotation.c
@@ -1,27 +1,17 @@
 #include <stdio.h>
-#include <stdlib.h>
 #include "rotation.h"
 
-int initProcess(rotationClient* client){
-	if( (client->process=popen("./host.py","r")) == NULL){
-		perror("could not execute host program");
-		return -1;
-	}else{
-		return 0;
-	}
+float getRotation(){
+	float rotation = -810;
+    printf("rotation:");
+	scanf("%f", &rotation);
+	printf("ok\n");
+	//値を抽出(loop)
+//	while(rotation == -810){
+//      printf("rotation:");
+//		scanf("%f", &rotation);
+//		printf("ok\n");
+//	}
+    return rotation;
 }
 
-void setRotation(rotationClient* client){
-	client->rotation = -810;
-	char *buf;
-	//値を抽出
-	while(client->rotation == -810 || client->rotation != -1919){
-		fgets(buf, 20, client->process);
-		client->rotation = atof(buf);
-		printf("%s\n",buf);
-	}
-}
-
-void closeProcess(rotationClient* client){
-	pclose(client->process);
-}
