+++ jissennohito-server/rotation.h
@@ -1,8 +1,1 @@
-typedef struct rotationClient{
-	FILE* process;
-	float rotation;
-}rotationClient;
-
-int initProcess(rotationClient* client);
-void setRotation(rotationClient* client);
-void closeProcess(rotationClient* client);
+float getRotation();
