+++ jissennohito-server/rotation2.h
@@ -1,0 +1,8 @@
+typedef struct rotationClient{
+	FILE* process;
+	float rotation;
+}rotationClient;
+
+int initProcess(rotationClient* client);
+void setRotation(rotationClient* client);
+void closeProcess(rotationClient* client);
