+++ jissennohito-server/rotationSample4.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include "rotation.h"
 #include "nctrb_mechctrl_io.h"
 
 #define ROTATE_CCW 0b01010101
@@ -6,11 +7,9 @@
 int main(int argc, char* argv[]){
 	float rotation = -810;
 	nctrb_mc_io_initialize();
-	while(rotation < 0){
+	while(rotation < 180){
 		Wheel(ROTATE_CCW);
-		printf("rotation:");
-		scanf("%f", &rotation);
-        printf("ok");
+		rotation = getRotation();
 	}
 	Wheel(0b11111111);
 	return 0;
